server/requirements.txt
flask==3.0.3
flask-cors==5.0.0
werkzeug==3.0.3
Flask-SQLAlchemy==3.0.4
python-dotenv==1.0.0
requests==2.32.3